src/requirements.txt
azure.data.tables
beautifulsoup4
load_dotenv
langchain-text-splitters
openai
requests
tqdm
